pyproject.toml
[tool.poetry]
name = "wesley-bot-whatsapp-assistant"
version = "0.1.0"
description = "Bot de WhatsApp (FastAPI + Evolution API)"
authors = ["Wesley"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.2.1"
httpx = ">=0.27.0"
alembic = "^1.13.1"
google-genai = "^1.64.0"
faiss-cpu = ">=1.7.4"
pypdf = "^6.7.1"
markdown = "^3.10.2"
numpy = "^2.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
